Update index.html
Updated title and started adding a back story. Since this is a fictional company I figured why not go for a fictional setting? There is nothing in the rubric that says the products have to be realistic, and I figured it would be a way to have some fun with the project.